// react-frontend/src/api/models/residence-entity.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Spys API
 * REST API for Spys
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * 
 * @export
 * @interface ResidenceEntity
 */
export interface ResidenceEntity {
    /**
     * 
     * @type {number}
     * @memberof ResidenceEntity
     */
    'id'?: number;
    /**
     * 
     * @type {string}
     * @memberof ResidenceEntity
     */
    'residence'?: string;
    /**
     * 
     * @type {string}
     * @memberof ResidenceEntity
     */
    'createdAt'?: string;
    /**
     * 
     * @type {string}
     * @memberof ResidenceEntity
     */
    'updatedAt'?: string;
}